requirements.txt
langchain
langchain_openai
PyTelegramBotAPI
python-dotenv